+++ packages/analytics-remote-config/rollup.config.js
@@ -1,6 +1,0 @@
-import { iife, umd } from '../../scripts/build/rollup.config';
-
-iife.input = umd.input;
-iife.output.name = 'remoteConfig';
-
-export default [umd, iife];