+++ packages/analytics-remote-config/src/index.ts
@@ -1,3 +1,0 @@
-import * as remoteConfigFetch from './remote-config';
-export { RemoteConfigFetch } from './types';
-export const { createRemoteConfigFetch } = remoteConfigFetch;