+++ packages/plugin-autocapture-browser/src/autocapture/track-rage-click.ts
@@ -160,7 +160,6 @@ export function trackRageClicks({
         }
 
         resetClickWindow(click);
-        return null;
       } else {
         clickWindow.push(click);
       }