+++ packages/plugin-autocapture-browser/test/autocapture-plugin/actions.test.ts
@@ -1,14 +1,17 @@
 import { autocapturePlugin } from '../../src/autocapture-plugin';
 import { type ElementBasedTimestampedEvent } from '../../src/helpers';
 
-import { BrowserConfig, EnrichmentPlugin, ILogger } from '@amplitude/analytics-core';
-import { createInstance } from '@amplitude/analytics-browser';
-import type {
+import {
   ElementInteractionsOptions,
   LabeledEvent,
   Trigger,
   DataSource,
-} from '@amplitude/analytics-core/lib/esm/types/element-interactions';
+  BrowserConfig,
+  EnrichmentPlugin,
+  ILogger,
+  IConfig,
+} from '@amplitude/analytics-core';
+import { createInstance } from '@amplitude/analytics-browser';
 import { getDataSource, executeActions } from '../../src/pageActions/actions';
 import { DataExtractor } from '../../src/data-extractor';
 
@@ -189,7 +192,7 @@ describe('page actions', () => {
       };
 
       plugin = autocapturePlugin(autocaptureConfig);
-      await plugin?.setup?.(config as BrowserConfig, instance);
+      await plugin?.setup?.(config as IConfig, instance);
 
       const button1 = document.querySelector('#product-card-1 .add-to-cart-button');
 
@@ -244,7 +247,7 @@ describe('page actions', () => {
       };
 
       plugin = autocapturePlugin(autocaptureConfig);
-      await plugin?.setup?.(config as BrowserConfig, instance);
+      await plugin?.setup?.(config as IConfig, instance);
 
       const button3 = document.querySelector('#product-card-3 .add-to-cart-button');
 
@@ -276,7 +279,7 @@ describe('page actions', () => {
       };
 
       plugin = autocapturePlugin(autocaptureConfig);
-      await plugin?.setup?.(config as BrowserConfig, instance);
+      await plugin?.setup?.(config as IConfig, instance);
 
       const button1 = document.querySelector('#product-card-1 .add-to-cart-button');