+++ packages/gtm-snippet/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.30.0](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/gtm-snippet@2.28.1...@amplitude/gtm-snippet@2.30.0) (2025-10-29)
+
+**Note:** Version bump only for package @amplitude/gtm-snippet
+
+
+
+
+
+# [2.29.0](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/gtm-snippet@2.27.2...@amplitude/gtm-snippet@2.29.0) (2025-10-23)
+
+**Note:** Version bump only for package @amplitude/gtm-snippet
+
+
+
+
+
+# [2.28.0](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/gtm-snippet@2.27.1...@amplitude/gtm-snippet@2.28.0) (2025-10-23)
+
+**Note:** Version bump only for package @amplitude/gtm-snippet
+
+
+
+
+
+## [2.27.1](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/gtm-snippet@2.26.3...@amplitude/gtm-snippet@2.27.1) (2025-10-21)
+
+**Note:** Version bump only for package @amplitude/gtm-snippet
+
+
+
+
+
 # [2.27.0](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/gtm-snippet@2.26.2...@amplitude/gtm-snippet@2.27.0) (2025-10-17)
 
 **Note:** Version bump only for package @amplitude/gtm-snippet